+++ src/main/java/com/umc/sopi/web/dto/UserBloodResponse.java
@@ -44,6 +44,6 @@ public class UserBloodResponse {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class DayDTO {
-        Integer LeftDay;
+        Long LeftDay;
     }
 }
